+++ src/blueetl/features.py
@@ -8,12 +8,10 @@ import pandas as pd
 
 from blueetl.cache import CacheManager
 from blueetl.constants import (
-    CIRCUIT,
     CIRCUIT_ID,
     GID,
     NEURON_CLASS,
     NEURON_CLASS_INDEX,
-    SIMULATION,
     SIMULATION_ID,
     TRIAL,
     WINDOW,
@@ -268,50 +266,64 @@ def calculate_features_multi(
 
 
 def call_by_simulation(
-    repo: Repository,
+    simulations: pd.DataFrame,
+    dataframes_to_filter: Dict[str, pd.DataFrame],
     func: Callable,
     jobs: Optional[int] = None,
     backend: Optional[str] = None,
+    how: str = "namedtuple",
 ) -> List[Any]:
     """Execute the given function in parallel, one task for each simulation.
 
     Args:
-        repo: repository instance.
-        func: callable called for each simulation, accepting
-            simulation_row: NamedTuple
-            simulation_spikes: pd.DataFrame
-            simulation_windows: pd.DataFrame
-            circuit_neurons: pd.DataFrame
-            circuit_neuron_classes: pd.DataFrame
+        simulations: DataFrame of simulations.
+        dataframes_to_filter: dict of DataFrames to filter by simulation_id and/or circuit_id,
+            and passed to each subprocess.
+        func: callable called for each simulation, accepting:
+            - simulation_row: NamedTuple (or the type specified with the `how` parameter)
+            - filtered_dataframes: dict of DataFrames filtered by simulation_id and/or circuit_id
+            If the function has other parameters, they can be applied using `functools.partials`,
+            and they will be serialized and passed unchanged to the subprocesses.
         jobs: number of jobs (see run_parallel)
         backend: parallel backend (see run_parallel)
+        how: format the `simulation_row` parameter passed to the func callback.
+            It can be one of "namedtuple", "dict", "series", "dataframe".
 
     Returns:
         list of results
     """
 
     def tasks_generator() -> Iterator[Task]:
-        for circuit_id in repo.simulations.df[CIRCUIT_ID].unique():
-            circuit_neurons = repo.neurons.df.etl.q(circuit_id=circuit_id)
-            circuit_neuron_classes = repo.neuron_classes.df.etl.q(circuit_id=circuit_id)
-            it = (
-                repo.simulations.df.drop(columns=[CIRCUIT, SIMULATION])
-                .etl.q(circuit_id=circuit_id)
-                .etl.iter()
-            )
+        # cache the names to avoid looking into df.index.names and df.columns in each iteration
+        cached_names = {
+            name: set(df).union(df.index.names) for name, df in dataframes_to_filter.items()
+        }
+        for circuit_id in simulations[CIRCUIT_ID].unique():
+            filtered_by_circuit_id = {
+                name: df.etl.q(circuit_id=circuit_id) if CIRCUIT_ID in cached_names[name] else df
+                for name, df in dataframes_to_filter.items()
+            }
+            it = simulations.etl.q(circuit_id=circuit_id).etl.iter()
             for _, simulation_row in it:
                 simulation_id = simulation_row.simulation_id
-                simulation_spikes = repo.spikes.df.etl.q(simulation_id=simulation_id)
-                simulation_windows = repo.windows.df.etl.q(simulation_id=simulation_id)
+                filtered_by_circuit_id_and_simulation_id = {
+                    name: df.etl.q(simulation_id=simulation_id)
+                    if SIMULATION_ID in cached_names[name]
+                    else df
+                    for name, df in filtered_by_circuit_id.items()
+                }
                 yield Task(
                     partial(
                         func,
-                        simulation_row=simulation_row,
-                        simulation_spikes=simulation_spikes,
-                        simulation_windows=simulation_windows,
-                        circuit_neurons=circuit_neurons,
-                        circuit_neuron_classes=circuit_neuron_classes,
+                        simulation_row=convert_row(simulation_row),
+                        filtered_dataframes=filtered_by_circuit_id_and_simulation_id,
                     )
                 )
 
+    convert_row = {
+        "namedtuple": lambda row: row,
+        "dict": lambda row: row._asdict(),
+        "series": lambda row: pd.Series(row._asdict()),
+        "dataframe": lambda row: pd.DataFrame([row]),
+    }[how]
     return run_parallel(tasks_generator(), jobs=jobs, backend=backend)